Create a proper AsyncGenerator in db streams
