+++ packages/db/src/controller/index.ts
@@ -3,4 +3,4 @@
  */
 
 export {IDatabaseController, IFilterOptions, IKeyValue} from "./interface";
-export {LevelDbController} from "./impl/level";
+export {LevelDbController} from "./level";
