+++ packages/lodestar/package.json
@@ -74,7 +74,6 @@
     "interface-datastore": "^2.0.0",
     "it-all": "^1.0.2",
     "it-pipe": "^1.1.0",
-    "it-pushable": "^1.4.0",
     "libp2p": "^0.30.2",
     "libp2p-bootstrap": "^0.12.1",
     "libp2p-gossipsub": "^0.8.0",
