+++ packages/lodestar/test/unit/api/rest/events/getEventStream.test.ts
@@ -2,7 +2,6 @@ import EventSource from "eventsource";
 import {URL} from "url";
 import {BeaconEvent, BeaconEventType} from "../../../../../src/api/impl/events";
 import {RestApi} from "../../../../../src/api/rest";
-import pushable from "it-pushable/index";
 import {generateAttestation} from "../../../../utils/attestation";
 import {expect} from "chai";
 import {AddressInfo} from "net";
@@ -16,10 +15,18 @@ describe("rest - events - getEventStream", function () {
     const restApi = await setupRestApiTestServer();
     const eventsApiStub = restApi.server.api.events as SinonStubbedInstance<EventsApi>;
 
-    const source = pushable<BeaconEvent>();
-    // @ts-ignore
-    source.stop = () => null;
-    eventsApiStub.getEventStream.returns((source as unknown) as LodestarEventIterator<BeaconEvent>);
+    const mockEventStream = new LodestarEventIterator<BeaconEvent>(({push}) => {
+      push({
+        type: BeaconEventType.BLOCK,
+        message: {slot: 1, block: Buffer.alloc(32, 0)},
+      });
+      push({
+        type: BeaconEventType.ATTESTATION,
+        message: generateAttestation(),
+      });
+    });
+
+    eventsApiStub.getEventStream.returns(mockEventStream);
     const eventSource = new EventSource(
       getEventStreamUrl([BeaconEventType.BLOCK, BeaconEventType.ATTESTATION], restApi)
     );
@@ -29,17 +36,7 @@ describe("rest - events - getEventStream", function () {
     const attestationEventPromise = new Promise((resolve) => {
       eventSource.addEventListener(BeaconEventType.ATTESTATION, resolve);
     });
-    source.push({
-      type: BeaconEventType.BLOCK,
-      message: {
-        slot: 1,
-        block: Buffer.alloc(32, 0),
-      },
-    });
-    source.push({
-      type: BeaconEventType.ATTESTATION,
-      message: generateAttestation(),
-    });
+
     const blockEvent = await blockEventPromise;
     const attestationEvent = await attestationEventPromise;
     expect(blockEvent).to.not.be.null;
